pyproject.toml
[project]
name = "sdr-research-agent"
version = "0.1.0"
description = "Automated SDR Research Agent - Multi-agent system for personalized cold email generation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "langgraph>=0.0.40",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.20",
    "pgvector>=0.2.4",
    "sqlalchemy>=2.0.25",
    "redis>=5.0.0",
    "firecrawl-py>=0.0.16",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.26.0",
    "helicone>=0.1.0",
    "streamlit>=1.29.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
